requirements.txt
dj-database-url==0.5.0
django==2.2
django-crispy-forms
gunicorn==19.9.0
psycopg2==2.7.5
pytz==2018.5
whitenoise==4.0
django-taggit==0.22.2
Pillow==7.0.0


